# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: ReplyX API
  version: 0.1.0
  description: >
    Staff and customer API of the ReplyX multi-tenant helpdesk. The tenant always comes from the host. Authentication
    uses the `rx_session` cookie, and state-changing requests must send an `X-CSRF-Token` header matching the `rx_csrf`
    cookie. Lists are cursor-paginated and wrapped as `{ "items": [...], "nextCursor": "..." | null }`. Creating POSTs
    accept an optional `Idempotency-Key` header (kept 24 h).
servers:
  - url: https://{tenant}.replyx.app/api/v1
    description: Tenant host
    variables:
      tenant:
        default: acme
        description: Tenant slug (subdomain)
  - url: https://console.replyx.app/api/v1
    description: Console host (platform operators)
security:
  - session: []
...
paths:
...
  /tickets/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      operationId: getTicket
      tags:
        - tickets
      summary: Ticket with properties, links and customer summary (ticket.view on the ticket's group)
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTicket
      tags:
        - tickets
      summary: Change title, state, priority, group, owner or tags (ticket.edit on the current group)
      description: >
        Only the fields sent are changed. A group change needs ticket.create on the destination,
        which must be active; from Ungrouped, ticket.edit on Ungrouped is enough (FR-041). ownerId
        must have edit on the (new) group (FR-040). pendingUntil is required to enter or remain in
        a pending state. State changes go through the ticket state machine.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                state:
                  $ref: "#/components/schemas/TicketState"
                pendingUntil:
                  type:
                    - string
                    - "null"
                  format: date-time
                priority:
                  $ref: "#/components/schemas/Priority"
                groupId:
                  type:
                    - string
                    - "null"
                  format: uuid
                ownerId:
                  type:
                    - string
                    - "null"
                  format: uuid
                tagIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: OWNER_NOT_ELIGIBLE, GROUP_INACTIVE or INVALID_TRANSITION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTicket
      tags:
        - tickets
      summary: Delete a ticket permanently (ticket.delete on its group)
      description: A hard delete; the customer's thread loses this ticket's messages. Writes a ticket.deleted audit entry.
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: rx_session
      description: Session cookie set on sign-in
    operatorSession:
      type: apiKey
      in: cookie
      name: rx_op_session
      description: Operator session cookie, set by platform sign-in on the console host
    supportToken:
      type: apiKey
      in: header
      name: X-Support-Token
      description: Read-only operator access to a tenant under an active support-access grant (FR-001a). Sent on the tenant
        host; every request is audited and anything but GET returns 403 READ_ONLY_SUPPORT_ACCESS.
  parameters:
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Id"
...
  responses:
    ValidationFailed:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthenticated:
      description: No valid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PermissionDenied:
      description: Resource is visible but the action is not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Missing, belongs to another tenant, or not visible to the caller
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Id:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required:
        - error
      additionalProperties: false
      properties:
        error:
          type: object
          required:
            - code
            - message
          additionalProperties: false
          properties:
            code:
              type: string
              pattern: ^[A-Z][A-Z0-9_]+$
              examples:
                - TICKET_NOT_FOUND
            message:
              type: string
            details:
              type: array
              description: Only present for validation errors
              items:
                type: object
                required:
                  - path
                  - issue
                properties:
                  path:
                    type: string
                  issue:
                    type: string
            retryAfter:
              type: integer
              description: Seconds, only for RATE_LIMITED
...
    TicketState:
      type: string
      enum:
        - new
        - open
        - pending_reminder
        - pending_close
        - resolved
        - closed
    Priority:
      type: string
      enum:
        - low
        - normal
        - high
        - urgent
    WaitingOn:
      type: string
      enum:
        - support
        - customer
...
    UserRef:
      type: object
      required:
        - id
        - name
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          type: string
        avatarUrl:
          type:
            - string
            - "null"
          format: uri
    GroupRef:
      type: object
      required:
        - id
        - name
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          type: string
    TagRef:
      type: object
      required:
        - id
        - name
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          type: string
...
    SlaSummary:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - none
            - ok
            - warning
            - breached
        policyName:
          type: string
        nextDueAt:
          type:
            - string
            - "null"
          format: date-time
        target:
          type: string
          enum:
            - first_response
            - next_response
            - resolution
    TicketSummary:
      type: object
      required:
        - id
        - number
        - title
        - state
        - priority
        - waitingOn
        - group
        - owner
        - customer
        - tags
        - sla
        - lastCustomerMessageAt
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/Id"
        number:
          type: integer
        title:
          type: string
        state:
          $ref: "#/components/schemas/TicketState"
        priority:
          $ref: "#/components/schemas/Priority"
        waitingOn:
          $ref: "#/components/schemas/WaitingOn"
        group:
          oneOf:
            - $ref: "#/components/schemas/GroupRef"
            - type: "null"
        owner:
          oneOf:
            - $ref: "#/components/schemas/UserRef"
            - type: "null"
        customer:
          $ref: "#/components/schemas/UserRef"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagRef"
        sla:
          $ref: "#/components/schemas/SlaSummary"
        lastCustomerMessageAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TicketLink:
      type: object
      required:
        - id
        - kind
        - direction
        - ticket
        - removedReason
      properties:
        id:
          $ref: "#/components/schemas/Id"
        kind:
          type: string
          enum:
            - follow_up_of
            - related
            - duplicate_of
            - merged_into
            - split_from
        direction:
          type: string
          enum:
            - outgoing
            - incoming
        ticket:
          description: null when the linked ticket was removed by retention or is not visible to the caller
          oneOf:
            - type: object
              properties:
                id:
                  type: string
                number:
                  type: integer
                title:
                  type: string
                state:
                  $ref: "#/components/schemas/TicketState"
            - type: "null"
        removedReason:
          type:
            - string
            - "null"
          enum:
            - retention
            - not_visible
            - null
    Ticket:
      allOf:
        - $ref: "#/components/schemas/TicketSummary"
        - type: object
          required:
            - pendingUntil
            - autoCloseAt
            - resolvedAt
            - closedAt
            - lastAgentReplyAt
            - origin
            - mergedInto
            - links
            - csat
            - allowedActions
          properties:
            pendingUntil:
              type:
                - string
                - "null"
              format: date-time
            autoCloseAt:
              type:
                - string
                - "null"
              format: date-time
            resolvedAt:
              type:
                - string
                - "null"
              format: date-time
            closedAt:
              type:
                - string
                - "null"
              format: date-time
            lastAgentReplyAt:
              type:
                - string
                - "null"
              format: date-time
            origin:
              type: string
              enum:
                - customer_message
                - staff_started
                - split
                - follow_up
            mergedInto:
              type:
                - string
                - "null"
              format: uuid
            links:
              type: array
              items:
                $ref: "#/components/schemas/TicketLink"
            csat:
              type:
                - object
                - "null"
              properties:
                score:
                  type: integer
                comment:
                  type: string
            allowedActions:
              type: array
              description: What the caller may do now (drives UI; server re-checks)
              items:
                type: string
                enum:
                  - reply
                  - note
                  - edit
                  - change_group
                  - assign
                  - delete
                  - merge
                  - split
                  - move_message
...